+++ java-client/src/test/java/org/opensearch/client/opensearch/integTest/AbstractMultiSearchRequestIT.java
@@ -9,9 +9,13 @@
 package org.opensearch.client.opensearch.integTest;
 
 import org.junit.Test;
+import org.opensearch.client.json.JsonData;
 import org.opensearch.client.opensearch._types.FieldSort;
 import org.opensearch.client.opensearch._types.FieldValue;
+import org.opensearch.client.opensearch._types.InlineScript;
 import org.opensearch.client.opensearch._types.Refresh;
+import org.opensearch.client.opensearch._types.Script;
+import org.opensearch.client.opensearch._types.ScriptField;
 import org.opensearch.client.opensearch._types.SortOptions;
 import org.opensearch.client.opensearch._types.SortOrder;
 import org.opensearch.client.opensearch._types.query_dsl.FuzzyQuery;
@@ -27,8 +31,13 @@ import org.opensearch.client.opensearch.core.search.Hit;
 import org.opensearch.client.opensearch.core.search.SourceConfig;
 import org.opensearch.client.util.ObjectBuilder;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.function.Function;
 
 public abstract class AbstractMultiSearchRequestIT extends OpenSearchJavaClientTestCase {
@@ -175,6 +184,44 @@ public abstract class AbstractMultiSearchRequestIT extends OpenSearchJavaClientT
 		assertEquals(1, response2.responses().get(0).result().hits().hits().size());
 	}
 
+	@Test
+	public void shouldReturnMultiSearchesScriptFields() throws Exception {
+		String index = "multiple_searches_request_script_fields";
+		createTestDocuments(index);
+
+		RequestItem sortedItemsQuery = createMSearchSortedFuzzyRequest();
+
+		MsearchResponse<ShopItem> response = sendMSearchRequest(index, List.of(sortedItemsQuery));
+		assertEquals(1, response.responses().size());
+		var hits = response.responses().get(0).result().hits().hits();
+		assertEquals(3, hits.size());
+		assertNull(hits.get(0).score());
+		assertNull(hits.get(1).score());
+		assertNull(hits.get(2).score());
+
+		Map<String, ScriptField> scriptFields = new HashMap<>();
+		scriptFields.put("quantity", new ScriptField.Builder().script(Script.of(s -> s.inline(new InlineScript.Builder()
+											.lang("painless")
+											.source("doc['quantity'].value + params.inc")
+                    						.params("inc", JsonData.of(1))
+											.build())))
+										.build());
+		
+
+		RequestItem requestItem = createMSearchQueryWithScriptFields("small", scriptFields);
+		
+		MsearchResponse<ShopItem> responseWithScriptFields = sendMSearchRequest(index, List.of(requestItem));
+		var hitsWithScriptFields = responseWithScriptFields.responses().get(0).result().hits().hits();
+		assertEquals(2, hitsWithScriptFields.size());
+		// validating that the quantity for small items is increased by 1
+    	ObjectMapper mapper = new ObjectMapper();
+		JsonNode node = mapper.readTree(hitsWithScriptFields.get(0).fields().get("quantity").toString());
+		assertEquals(2, (int) mapper.treeToValue(node.get(0), int.class));
+		
+		node = mapper.readTree(hitsWithScriptFields.get(1).fields().get("quantity").toString());
+		assertEquals(3, (int) mapper.treeToValue(node.get(0), int.class));
+	}
+
 
 	private void assertResponseSources(MultiSearchResponseItem<ShopItem> response) {
 		List<Hit<ShopItem>> hitsWithHighlights = response.result().hits().hits();
@@ -196,6 +243,14 @@ public abstract class AbstractMultiSearchRequestIT extends OpenSearchJavaClientT
 
 	private RequestItem createMSearchQuery(String itemSize) {
 		return createMSearchQuery(itemSize, null, List.of());
+	}
+
+	private RequestItem createMSearchQueryWithScriptFields(String itemSize, Map<String, ScriptField> scriptFields) {
+		return RequestItem.of(item -> item.header(header -> header)
+				.body(body -> body.query(createItemSizeSearchQuery(itemSize))
+								.scriptFields(scriptFields)
+						)
+				);
 	}
 
 	private RequestItem createMSearchQueryWithHighlight(String itemSize) {